requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
pydantic[email-validator]==2.10.0
redis==5.2.0
alembic==1.14.0
python-dotenv==1.0.1
